Change run_step2.sh to two parallel tasks.
run_step2.sh used to run four instances of step2.py in parallel, but
step2.py itself uses four processes to copy and skim ntuples. Perhaps
this was doing bad things on lxplus, which is why jobs would finish
but some of the ntuples would be corrupted. Or maybe it was a memory
issue with too many things open. Anyhow, this corruption error did not
appear in the step2.log file after setting run_step2.sh to run just two
two parallel instances of step2.py.