Fix paralleization bug in step2.py, add skim.py.
step2.py was not copying and skimming the Step1 ntuples in parallel
because I placed the apply_async calls within a list, so they were
actually called sequentially (print statements gave the bug away).

The code used for skimming is generally applicable to all steps, so
standardize it and make it its own function.